fix: utilizando os advices

+++ src/main/java/br/com/ufmt/backendsgccagendamento/services/ConsultaService.java
@@ -4,10 +4,11 @@ import br.com.ufmt.backendsgccagendamento.dtos.ConsultaDTO;
 import br.com.ufmt.backendsgccagendamento.entities.Consulta;
 import br.com.ufmt.backendsgccagendamento.entities.Medico;
 import br.com.ufmt.backendsgccagendamento.entities.Pessoa;
+import br.com.ufmt.backendsgccagendamento.exceptions.BadRequestException;
+import br.com.ufmt.backendsgccagendamento.exceptions.EntityNotFoundException;
 import br.com.ufmt.backendsgccagendamento.repositories.ConsultaRepository;
 import br.com.ufmt.backendsgccagendamento.repositories.MedicoRepository;
 import br.com.ufmt.backendsgccagendamento.repositories.PessoaRepository;
-import lombok.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -32,23 +33,24 @@ public class ConsultaService {
     }
 
     public Consulta buscarConsultaPorId(UUID id) {
-        return consultaRepository.findById(id).orElse(null);
+        return consultaRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(Consulta.class, id));
     }
 
     @Transactional
     public Consulta agendarConsulta(ConsultaDTO consultaDTO) {
         if (consultaRepository.existsByMedicoAndData(consultaDTO.getMedicoId(), consultaDTO.getData())) {
-            throw new RuntimeException("Horário indisponível para este médico.");
+            throw new BadRequestException("Horário indisponível para este médico.");
         }
 
         if (consultaRepository.existsByPacienteAndData(consultaDTO.getPacienteId(), consultaDTO.getData())) {
-            throw new RuntimeException("Paciente já possui uma consulta neste horário.");
+            throw new BadRequestException("Paciente já possui uma consulta neste horário.");
         }
 
         Medico medico = medicoRepository.findById(consultaDTO.getMedicoId())
-                .orElseThrow(() -> new RuntimeException("Médico não encontrado"));
+                .orElseThrow(() -> new EntityNotFoundException(Medico.class, consultaDTO.getMedicoId()));
         Pessoa paciente = pessoaRepository.findById(consultaDTO.getPacienteId())
-                .orElseThrow(() -> new RuntimeException("Paciente não encontrado"));
+                .orElseThrow(() -> new EntityNotFoundException(Pessoa.class, consultaDTO.getPacienteId()));
 
         Consulta novaConsulta = new Consulta();
         novaConsulta.setMedico(medico);
@@ -59,6 +61,9 @@ public class ConsultaService {
 
     @Transactional
     public void cancelarConsulta(UUID id) {
+        if(!consultaRepository.existsById(id)){
+            throw new EntityNotFoundException(Consulta.class, id);
+        }
         consultaRepository.deleteById(id);
     }
 }
